refactor(spec): keep the editability-boundary options package-internal (#7887)
`check:api-surface` (repo-wide TypeScript Type Check job) went red: exporting
`VISIBILITY_ONLY_STRICT_OPTIONS` from `shared/visibility.ts` put it in the
public barrel and moved the package's API surface.

Moved to a new `shared/editability-boundary.ts`, which the barrel deliberately
does not re-export — the same posture as `strict-object.ts` and
`alias-probe.ts`. `StrictObjectOptions`, the const's own type, is not public
either, so a published value of that type is one no consumer could annotate.
The public API surface and `export-origins` now both read unchanged, which is
also the card's own claim one level out.
